# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Limehome API",
    "description": "Limehome API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/property": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "at",
            "type": "string",
            "required": true,
            "description": "Comma separated point coordinates (Latitude and Longitude)"
          }
        ],
        "tags": [
          "Property"
        ],
        "summary": "Get all the properties",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Properties"
            }
          }
        }
      }
    },
    "/api/property/{id}/bookings": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "type": "integer",
            "name": "id",
            "minimum": 1,
            "required": true,
            "description": "Property ID"
          }
        ],
        "tags": [
          "Property"
        ],
        "summary": "Get all the bookings of the properties",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Bookings"
            }
          }
        }
      }
    },
    "/api/bookings": {
      "put": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Booking"
            }
          }
        ],
        "tags": [
          "Booking"
        ],
        "summary": "Create booking of the property",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Booking"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Property": {
      "required": [
        "property_id",
        "title",
        "vicinity",
        "lat_lon"
      ],
      "properties": {
        "property_id": {
          "type": "integer",
          "uniqueItems": true
        },
        "title": {
          "type": "string"
        },
        "vicinity": {
          "type": "string"
        },
        "lat_lon": {
          "type": "string"
        }
      }
    },
    "Properties": {
      "type": "array",
      "$ref": "#/definitions/Property"
    },
    "Booking": {
      "required": [
        "booking_id",
        "property_id",
        "customer",
        "arrival_date",
        "departure_date",
        "price",
        "guests"
      ],
      "properties": {
        "booking_id": {
          "type": "number"
        },
        "property_id": {
          "type": "number"
        },
        "customer": {
          "type": "string"
        },
        "arrival_date": {
          "type": "string"
        },
        "departure_date": {
          "type": "string"
        },
        "price": {
          "type": "string"
        },
        "guests": {
          "type": "number"
        }
      }
    },
    "Bookings": {
      "type": "array",
      "$ref": "#/definitions/Booking"
    }
  }
}